AP-7241 bugfixes in DurationDeserializer and DurationSerializer regarding pyhton3 compatibility (now working in python2 and 3), bugfix in Flatbuffers: there was an error when serializing BYTES datatype without specifying a serializer, now BYTES may be sent without a serializer,too. hotfix in Flatbuffers: on receiving BYTES type value is converted to range (0,256) (NOTE: this is not a bugfix, just a workaround for the time being)

+++ org.knime.python2/src/org/knime/python2/typeextension/builtin/duration/DurationDeserializer.py
@@ -5,7 +5,7 @@ def deserialize(bytes):
     hidx = durationstr.find('H');
     hours = 0
     if hidx >= 0:
-        hours = long(durationstr[:hidx])
+        hours = int(durationstr[:hidx])
         durationstr = durationstr[hidx+1:]
     midx = durationstr.find('M')
     minutes = 0
